+++ voicenavigator-20180612/alibabacloud_voicenavigator20180612/models/_modify_ttsconfig_request.py
@@ -9,6 +9,7 @@ class ModifyTTSConfigRequest(DaraModel):
         self,
         ali_customized_voice: str = None,
         app_key: str = None,
+        background_music_name: str = None,
         engine: str = None,
         engine_xunfei: str = None,
         ext_params: str = None,
@@ -24,6 +25,7 @@ class ModifyTTSConfigRequest(DaraModel):
         self.ali_customized_voice = ali_customized_voice
         # The AppKey of the third-party voice configuration.
         self.app_key = app_key
+        self.background_music_name = background_music_name
         # The TTS engine.
         self.engine = engine
         # The iFLYTEK engine parameters.
@@ -59,6 +61,9 @@ class ModifyTTSConfigRequest(DaraModel):
 
         if self.app_key is not None:
             result['AppKey'] = self.app_key
+
+        if self.background_music_name is not None:
+            result['BackgroundMusicName'] = self.background_music_name
 
         if self.engine is not None:
             result['Engine'] = self.engine
@@ -99,6 +104,9 @@ class ModifyTTSConfigRequest(DaraModel):
 
         if m.get('AppKey') is not None:
             self.app_key = m.get('AppKey')
+
+        if m.get('BackgroundMusicName') is not None:
+            self.background_music_name = m.get('BackgroundMusicName')
 
         if m.get('Engine') is not None:
             self.engine = m.get('Engine')